Implement grunt-watch-exec
This commit moves the `grunt-watch-exec` code that was in the viper repo
and moves it into this repo. This commit also cleans up the grunt config
and changes the copyrights to Squaremouth instead of Gale. This commit
also breaks the tests because they were the generated tests and not
tests specific to the `grunt-watch-exec` plugin. We should clean up and
create valid tests for this plugin.

[38469757361489](https://app.asana.com/0/38469757361489/38469757361489)